requirements.txt
asgiref==3.5.2
click==8.1.3
cmake==3.23.3
colorama==0.4.5
cycler==0.11.0
Cython==0.29.28
Django==4.1
django-crispy-forms==1.14.0
django-extra-keywords==1.1.2
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.34.4
gensim==4.2.0
imageio==2.21.0
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.2
mysql==0.0.3
mysqlclient==2.1.1
networkx==2.8.5
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.0
six==1.16.0
sklearn==0.0
smart-open==6.0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tifffile==2022.8.3
tzdata==2022.1
